Add multiargs option to store multiple args in an array
Make it possible to collect the arguments to options in an array by
setting `optarg.multiargs=true` so multiple occurances of options with
args can be collected.

For example, when optarg.multiargs=true `... -p foo -p bar` can be
collected into `p = {'foo', 'bar'}`

By using a separate flag for this we keep backwards compatibility.

ref https://github.com/ncopa/lua-optarg/pull/3